requirements.txt
flask
flask-cors
mongomock
pymongo
pymupdf
python-docx
python-dotenv
numpy
requests
